rename: image service 구현체 이름 수정
