+++ src/main/java/com/three/ott_suggestion/image/service/PostImageServiceImpl.java
@@ -1,12 +1,12 @@
 package com.three.ott_suggestion.image.service;
 
-
 import com.three.ott_suggestion.global.exception.InvalidInputException;
-import com.three.ott_suggestion.image.entity.UserImage;
-import com.three.ott_suggestion.image.repository.UserImageRepository;
-import com.three.ott_suggestion.user.entity.User;
+import com.three.ott_suggestion.image.entity.PostImage;
+import com.three.ott_suggestion.image.repository.PostImageRepository;
+import com.three.ott_suggestion.post.entity.Post;
 import java.io.File;
 import java.io.IOException;
+import java.net.MalformedURLException;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
@@ -14,52 +14,49 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 
-
 @Service
 @RequiredArgsConstructor
 @Transactional(readOnly = true)
-public class UserImageService implements ImageService<UserImage> {
-
-    private final UserImageRepository userImageRepository;
+public class PostImageServiceImpl implements ImageService<PostImage> {
 
     @Value("${upload.path}")
     private String uploadPath;
 
+    private final PostImageRepository postImageRepository;
+
     @Override
     @Transactional
-    public UserImage createImage(MultipartFile imageFile) throws Exception {
-        UserImage image = getUserImage(imageFile);
-        userImageRepository.save(image);
+    public PostImage createImage(MultipartFile file) throws Exception {
+        PostImage image = getPostImage(file);
+        postImageRepository.save(image);
         return image;
     }
 
     @Override
-    public String getImage(Long id) {
-        UserImage image = userImageRepository.findById(id)
-            .orElseThrow(() -> new InvalidInputException("프로필 이미지가 존재하지 않습니다."));
+    public String getImage(Long id) throws MalformedURLException {
+        PostImage image = postImageRepository.findById(id)
+            .orElseThrow(() -> new InvalidInputException("게시물 이미지가 존재하지 않습니다"));
         return image.getFilePath();
     }
 
-
     @Transactional
     @Override
-    public void updateImage(User user, MultipartFile imageFile) throws IOException {
-        UserImage image = getUserImage(imageFile);
-        UserImage userImage = userImageRepository.findById(user.getUserImage().getId())
-            .orElseThrow(() -> new InvalidInputException("프로필 이미지가 존재하지 않습니다."));
-        userImage.updateUserImage(image);
+    public void updateImage(Post post, MultipartFile imageFile) throws IOException {
+        PostImage image = getPostImage(imageFile);
+        PostImage postImage = postImageRepository.findById(post.getPostImage().getId())
+            .orElseThrow(() -> new InvalidInputException("게시물 이미지가 존재하지 않습니다"));
+        postImage.updatePostImage(image);
     }
 
-    private UserImage getUserImage(MultipartFile imageFile) throws IOException {
-        String originalFilename = imageFile.getOriginalFilename();
+    private PostImage getPostImage(MultipartFile file) throws IOException {
+        String originalFilename = file.getOriginalFilename();
         String saveFileName = createSaveFileName(originalFilename);
-        imageFile.transferTo(new File(getFullPath(saveFileName)));
-
+        file.transferTo(new File(getFullPath(saveFileName)));
         String filePath = uploadPath + saveFileName;
 
-        String contentType = imageFile.getContentType();
+        String contentType = file.getContentType();
 
-        UserImage image = UserImage.builder()
+        PostImage image = PostImage.builder()
             .fileName(originalFilename)
             .saveFileName(saveFileName)
             .contentType(contentType)
@@ -67,7 +64,6 @@ public class UserImageService implements ImageService<UserImage> {
             .build();
         return image;
     }
-
 
     private String createSaveFileName(String originalFilename) {
         String ext = extractExt(originalFilename);
@@ -83,4 +79,6 @@ public class UserImageService implements ImageService<UserImage> {
     private String getFullPath(String filename) {
         return uploadPath + filename;
     }
+
+
 }
